Extract category customization helpers
