+++ tests/test-audit.js
@@ -64,6 +64,7 @@ console.log('3. XSS Prevention');
 const viewsSrc = read('js/views.js');
 const lensPageShellSrc = read('js/lens-page-shell.js');
 const categoryViewRenderersSrc = read('js/category-view-renderers.js');
+const categoryCustomizationSrc = read('js/category-customization.js');
 const focusCardSrc = read('js/focus-card.js');
 const compareCorrelationsSrc = read('js/compare-correlations.js');
 const markerDetailSrc = read('js/marker-detail-modal.js');
@@ -114,6 +115,11 @@ assert('renderFattyAcidsView returns "" on unsafe categoryKey',
   /export function renderFattyAcidsView[^{]*\{[\s\S]{0,400}if\s*\(\s*!safeMarkerId\(categoryKey\)\s*\)\s*return\s*''/.test(categoryViewRenderersSrc));
 assert('showCategory chart-cards loop skips legacy customMarkers with unsafe keys',
   /for\s*\(\s*const\s*\[\s*key\s*,\s*marker\s*\]\s+of\s+withData\s*\)\s*\{\s*[\s\S]{0,200}if\s*\(\s*!safeMarkerId\(key\)\s*\)\s*continue/.test(viewsSrc));
+assert('category-customization.js owns rename/icon helpers',
+  /export async function renameCategory/.test(categoryCustomizationSrc) &&
+  /export async function renameMarker/.test(categoryCustomizationSrc) &&
+  /export function changeCategoryIcon/.test(categoryCustomizationSrc) &&
+  /export function showEmojiPicker/.test(categoryCustomizationSrc));
 
 // ═══════════════════════════════════════
 // 3c. Sweep guard — every innerHTML site in production JS is sanitized
@@ -135,7 +141,7 @@ const _SAFE_HELPERS = new Set([
   // is the markdown.js sanitized full renderer)
   'escapeHTML', 'renderMarkdown',
 ]);
-const _SWEEP_FILES = ['views.js', 'category-view-renderers.js', 'focus-card.js', 'marker-detail-modal.js', 'dashboard-widget-renderers.js', 'light-conditions-now.js', 'light-sessions-view.js', 'compare-correlations.js', 'mobile-dashboard.js', 'chat.js', 'charts.js'];
+const _SWEEP_FILES = ['views.js', 'category-view-renderers.js', 'category-customization.js', 'focus-card.js', 'marker-detail-modal.js', 'dashboard-widget-renderers.js', 'light-conditions-now.js', 'light-sessions-view.js', 'compare-correlations.js', 'mobile-dashboard.js', 'chat.js', 'charts.js'];
 
 function _sweepInnerHTML(filename, src) {
   const lines = src.split('\n');
